+++ packages/db/src/schema/index.ts
@@ -21,6 +21,7 @@ export * from "./orientation";
 export * from "./payments";
 export * from "./product-events";
 export * from "./projects-tasks";
+export * from "./purchase-orders";
 export * from "./quickdash";
 export * from "./quickengine";
 export * from "./quotes-estimates";